+++ src/m_config.h
@@ -379,6 +379,8 @@ typedef enum
 
 #define vid_windowsize_default                  "768x480"
 
+#define wad_default                             ""
+
 #define weaponbob_min                           0
 #define weaponbob_default                       75
 #define weaponbob_max                           100