+++ releasenotes.md
@@ -3,6 +3,7 @@
 * Optimizations have been made to further improve the overall performance and stability of *DOOM Retro*, particularly in the handling of large numbers of blood splats.
 * *DOOM Retro’s* splash screen has been redesigned.
 * A bug has been fixed whereby in some rare instances objects would get suspended in midair when the sector they were on lowered.
+* The last WAD to be opened now appears in the `File name:` field in *DOOM Retro’s* WAD launcher.
 * The `r_shakescreen` CVAR has been renamed to `r_shake_damage` and is now `50%` by default.
 * The screen will now shake when the player is near enough to an exploding barrel or rocket. This feature may be disabled using the new `r_shake_explode` CVAR.
 * When the `vid_showfps` CVAR is `on`, the frames per second displayed will be amber in color if between 35 and 60.